Common file created for 2 functions and readme's 1 section done
