+++ exercise-4/.vscode/settings.json
@@ -1,0 +1,54 @@
+{
+  "files.associations": {
+    "iostream": "cpp",
+    "array": "c",
+    "atomic": "c",
+    "bit": "c",
+    "cctype": "c",
+    "charconv": "c",
+    "clocale": "c",
+    "cmath": "c",
+    "compare": "c",
+    "concepts": "c",
+    "cstdarg": "c",
+    "cstddef": "c",
+    "cstdint": "c",
+    "cstdio": "c",
+    "cstdlib": "c",
+    "ctime": "c",
+    "cwchar": "c",
+    "cwctype": "c",
+    "deque": "c",
+    "string": "c",
+    "unordered_map": "c",
+    "vector": "c",
+    "exception": "c",
+    "algorithm": "c",
+    "functional": "c",
+    "iterator": "c",
+    "memory": "c",
+    "memory_resource": "c",
+    "numeric": "c",
+    "optional": "c",
+    "random": "c",
+    "string_view": "c",
+    "system_error": "c",
+    "tuple": "c",
+    "type_traits": "c",
+    "utility": "c",
+    "format": "c",
+    "initializer_list": "c",
+    "iosfwd": "c",
+    "istream": "c",
+    "limits": "c",
+    "new": "c",
+    "numbers": "c",
+    "ostream": "c",
+    "span": "c",
+    "stdexcept": "c",
+    "streambuf": "c",
+    "typeinfo": "c",
+    "variant": "c",
+    "text_encoding": "c"
+  }
+}
