+++ exercise-4/.vscode/tasks.json
@@ -1,0 +1,39 @@
+{
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "label": "Build Client",
+      "type": "shell",
+      "command": "g++",
+      "args": [
+        "src/tcp_echo_client.cpp",
+        "src/shared.cpp",
+        "-o", "build/client",
+        "-std=c++20",
+        "-O2",
+        "-Wall",
+        "-g",
+        "-fsanitize=address"
+      ],
+      "group": "build",
+      "problemMatcher": []
+    },
+    {
+      "label": "Build Server",
+      "type": "shell",
+      "command": "g++",
+      "args": [
+        "src/tcp_echo_server.cpp",
+        "src/shared.cpp",
+        "-o", "build/server",
+        "-std=c++20",
+        "-O2",
+        "-Wall",
+        "-g",
+        "-fsanitize=address"
+      ],
+      "group": "build",
+      "problemMatcher": []
+    }
+  ]
+}
