+++ exercise-4/src/shared.cpp
@@ -1,0 +1,14 @@
+#include "shared.h"
+
+void check_error(bool test, std::string error_message) {
+  if (test) {
+    std::cerr << error_message << "\n";
+    exit(EXIT_FAILURE);
+  }
+}
+
+int create_socket() {
+  int sock = socket(AF_INET, SOCK_STREAM, 0);
+  check_error(sock < 0, "Socket creation error\n");
+  return sock;
+}
