+++ exercise-4/src/tcp_echo_client.cpp
@@ -6,19 +6,8 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include "shared.h"
 
-void check_error(bool test, std::string error_message) {
-  if (test) {
-    std::cerr << error_message << "\n";
-    exit(EXIT_FAILURE);
-  }
-}
-
-int create_socket() {
-  int sock = socket(AF_INET, SOCK_STREAM, 0);
-  check_error(sock < 0, "Socket creation error\n");
-  return sock;
-}
 
 sockaddr_in create_address(const std::string &server_ip, int port) {
   sockaddr_in address;
