+++ exercise-4/src/tcp_echo_server.cpp
@@ -3,20 +3,15 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <unistd.h>
+#include "shared.h"
+// template <typename T, typename S> void check_error(T test, S error_message) {
+//   if (test) {
+//     std::cerr << error_message << "\n";
+//     exit(EXIT_FAILURE);
+//   }
+// }
 
-template <typename T, typename S> void check_error(T test, S error_message) {
-  if (test) {
-    std::cerr << error_message << "\n";
-    exit(EXIT_FAILURE);
-  }
-}
 
-int create_socket() {
-  int my_sock;
-  my_sock = socket(AF_INET, SOCK_STREAM, 0);
-  check_error(my_sock < 0, "Socket creation error\n");
-  return my_sock;
-}
 
 void set_socket_options(int sock, int opt) {
   auto err_code = setsockopt(sock, SOL_SOCKET, SO_REUSEADDR | SO_REUSEPORT,
